refactor: Add state slice to track shelf state in parent component
Previously the app was not tracking shelf state locally - it was relying on updating and reading from the API
to determine which shelf a book was on.

However, the 'search' API route doesn't preserve shelf information.
To address this, this commit adds a 'shelf' object to App's state
and passes down that information to all child components.

This project now meets all the Udacity Grading Rubric.